Move blocked UAs to own file
